refactor api class
