+++ deneg_core/scripts/agent.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-from deneg import DeNeg
+from deneg import DeNeg, Evaluator
 import time
 import argparse
 import random
@@ -10,29 +10,30 @@ import random
 class Agent(DeNeg):
     def __init__(self, name):
         super().__init__(name)
+        self.name = name
     
     def receive_alert(self, id, content):
         print(f"   received alert with id {id}")
         return True
 
-    def proposal_submission(self, round):
+    def proposal_submission(self, id, round):
         cost = random.uniform(1.0, 10.0)
-        print(f" \n ======================================= \n \
-            [{self.name}] propose cost: [{cost}] \
-            \n ======================================= \n")
+        print(f" \n ======================================= \n"
+              f"{id}: [{self.name}] propose cost: [{cost}]"
+              "\n ======================================= \n")
         return {"cost": cost }
 
-    def concession(self, results):
-        print(f"   received results {results}")
+    def concession(self, id, results):
+        print(f"{id}: received results {results}")
         return True
 
-    # def round_table(self, round: int, other_proposals):
-    #     return []
+    def round_table(self, id, round: int, other_proposals):
+        return Evaluator.LowestCostEvaluater(other_proposals)
 
-    def assignment(self, assignment):
-        print(f" \n ======================================= \n \
-            [{self.name}] received assignment [{assignment}] \
-            \n ======================================= \n")
+    def assignment(self, id, assignment):
+        print(f" \n ======================================= \n"
+              f"{id}: [{self.name}] received assignemnt: [{assignment}]"
+              "\n ======================================= \n")
         return True
 
 def main(agent_name):
