+++ deneg_msgs/msg/Proposal.msg
@@ -8,8 +8,11 @@ string proponent
 # negotiation round
 int8 round
 
-# overall cost
+# cost of the new proposal
 float32 cost
+
+# current cost of the participant
+float32 current_cost
 
 # optional constraints info
 string constraints
